+++ data-pipeline/src/insert-data.ts
@@ -2,25 +2,10 @@ import { DataSource } from 'typeorm';
 
 import { databaseOptions } from './common/common.constants';
 import { ProcessedData } from './types/processed-data.type';
+import { DataSourceManager } from './data-source-manager';
 
-const createDataSource = async () => {
-  const dataSource: DataSource = new DataSource(databaseOptions);
-
-  try {
-    return dataSource.initialize();
-  } catch (e) {
-    console.table(e);
-    console.error(e);
-    if (e.code === 'ENOTFOUND') {
-      console.error('Please check the database connection options in the .env file.\n');
-    }
-    await dataSource.destroy();
-    process.exit(1);
-  }
-};
-
-const bulkInsert = async (AppDataSource: DataSource, data: ProcessedData[]) => {
-  const repository = AppDataSource.getRepository('Restaurant');
+const bulkInsert = async (dataSourceManager: DataSourceManager, data: ProcessedData[]) => {
+  const repository = dataSourceManager.getRepository('Restaurant');
   const CHUNK_SIZE = 5000;
 
   for (let i = 0; i < data.length; i += CHUNK_SIZE) {
@@ -33,16 +18,13 @@ const bulkInsert = async (AppDataSource: DataSource, data: ProcessedData[]) => {
 
 // FIXME: insertData
 const insertData = async (dataList: ProcessedData[]) => {
-  const dataSource = await createDataSource();
+  const dataSourceManager = DataSourceManager.getInstance();
 
-  try {
-    await bulkInsert(dataSource, dataList);
-  } catch (e) {
-    console.error(e);
-    await dataSource.destroy();
-    process.exit(1);
-  }
-  await dataSource.destroy();
+  await dataSourceManager.initialize();
+
+  await bulkInsert(dataSourceManager, dataList);
+
+  await dataSourceManager.closeConnection();
 };
 
-export { createDataSource, insertData };
+export { insertData };
